moving functions from header file to c file to avoid multiple definitions, left the inline functions since they get copied locally
